Move yank and unyank to version::yank::*
